Atualizando o status HTTP dos métodos de DELETE

+++ src/main/java/br/com/oceandex/controllers/AnimalController.java
@@ -84,7 +84,12 @@ public class AnimalController {
     @DeleteMapping("{id}")
     @Transactional
     public ResponseEntity<Void> deletar(@PathVariable("id") Long id) {
-        repository.deleteById(id);
-        return ResponseEntity.noContent().build();
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
